+++ lib/src/widgets/_impl/_month_picker.dart
@@ -199,6 +199,11 @@ class _MonthPickerState extends State<_MonthPicker> {
     assert(debugCheckHasMaterial(context));
     return Column(
       children: <Widget>[
+        Divider(
+          color: widget.config.hideMonthPickerDividers == true
+              ? Colors.transparent
+              : null,
+        ),
         Expanded(
           child: GridView.builder(
             controller: _scrollController,
@@ -208,6 +213,11 @@ class _MonthPickerState extends State<_MonthPicker> {
             padding:
                 const EdgeInsets.symmetric(horizontal: _monthPickerPadding),
           ),
+        ),
+        Divider(
+          color: widget.config.hideMonthPickerDividers == true
+              ? Colors.transparent
+              : null,
         ),
       ],
     );